[!] remove compilation warnings

+++ source/geometry/tesellation/delaunay/delaunay3D.cpp
@@ -586,7 +586,7 @@ namespace internal {
 
 #if _DEBUG
 		for( int i = 0; i < 4; i++ ) {
-			assert( srcT.v[ i ] != pointIndex );
+			assert( srcT.v[ i ] != (int)pointIndex );
 		}
 #endif
